move AVR port's timers to a different translation unit
this gives greater flexibility in terms of timer choices
